feat(conta.service + controller + routes + models): implementa rota conta/transacoes/:CodCliente
A rota retorna o histórico de transações do cliente. Foi modificado os models e migrations para
conseguir valores com duas casas decimais

+++ src/database/migrations/20220716230257-create-cliente.js
@@ -18,7 +18,7 @@ module.exports = {
       },
       Saldo: {
         allowNull: false,
-        type: Sequelize.FLOAT
+        type: Sequelize.FLOAT(20,2)
       },
     });
   },
